+++ src/transform.ts
@@ -11,6 +11,45 @@ export const ALGORAND_ZERO_ADDRESS = 'AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA
 export const ALGORAND_ZERO_ADDRESS_BYTES = algosdk.decodeAddress(ALGORAND_ZERO_ADDRESS).publicKey
 
 /**
+ * Gets the synthetic transaction for the block payout as defined in the indexer
+ *
+ * @see https://github.com/algorand/indexer/blob/084577338ad4882f5797b3e1b30b84718ad40333/idb/postgres/internal/writer/write_txn.go?plain=1#L180-L202
+ */
+export function getTransactionFromBlockPayout(block: algosdk.modelsv2.BlockResponse, getRoundOffset: () => number): TransactionInBlock {
+  const pay = new algosdk.Transaction({
+    type: algosdk.TransactionType.pay,
+    sender: block.block.header.rewardState.feeSink,
+    note: new Uint8Array(Buffer.from(`ProposerPayout for Round ${block.block.header.round}`)),
+    suggestedParams: {
+      firstValid: block.block.header.round,
+      lastValid: block.block.header.round,
+      genesisID: block.block.header.genesisID,
+      genesisHash: block.block.header.genesisHash,
+      fee: 0n,
+      minFee: 0n,
+    },
+    paymentParams: {
+      receiver: block.block.header.proposer,
+      amount: block.block.header.proposerPayout,
+    },
+  })
+
+  const txn: TransactionInBlock = {
+    transactionId: pay.txID(),
+    roundTimestamp: Number(block.block.header.timestamp),
+    transaction: pay,
+    intraRoundOffset: getRoundOffset(),
+    roundNumber: block.block.header.round,
+    signedTxnWithAD: new algosdk.SignedTxnWithAD({
+      applyData: new algosdk.ApplyData({}),
+      signedTxn: new algosdk.SignedTransaction({ txn: pay }),
+    }),
+  }
+
+  return txn
+}
+
+/**
  * Processes a block and returns all transactions from it, including inner transactions, with key information populated.
  * @param blockResponse An Algorand block
  * @returns An array of processed transactions from the block
@@ -21,7 +60,7 @@ export function getBlockTransactions(blockResponse: algosdk.modelsv2.BlockRespon
   let roundOffset = 0
   const getRoundOffset = () => roundOffset++
 
-  return (block.payset ?? []).flatMap((signedTransactionInBlock) => {
+  const txns = (block.payset ?? []).flatMap((signedTransactionInBlock) => {
     // Assume that Consensus.RequireGenesisHash is true, we always copy the genesis hash from the block to the transaction
     const genesisHash = Buffer.from(blockResponse.block.header.genesisHash)
     const genesisId = signedTransactionInBlock.hasGenesisID ? blockResponse.block.header.genesisID : undefined
@@ -49,6 +88,12 @@ export function getBlockTransactions(blockResponse: algosdk.modelsv2.BlockRespon
       ),
     ]
   })
+
+  if (blockResponse.block.header.proposerPayout > 0n) {
+    txns.push(getTransactionFromBlockPayout(blockResponse, getRoundOffset))
+  }
+
+  return txns
 }
 
 function getBlockInnerTransactions(
@@ -186,202 +231,202 @@ export function getIndexerTransactionFromAlgodTransaction(t: TransactionInBlock,
       filtersMatched: filterName ? [filterName] : undefined,
       ...(transaction.type === TransactionType.acfg && transaction.assetConfig
         ? {
-            assetConfigTransaction: new algosdk.indexerModels.TransactionAssetConfig({
-              assetId: transaction.assetConfig.assetIndex,
-              params: createdAssetId
+          assetConfigTransaction: new algosdk.indexerModels.TransactionAssetConfig({
+            assetId: transaction.assetConfig.assetIndex,
+            params: createdAssetId
+              ? new algosdk.indexerModels.AssetParams({
+                creator: transaction.sender.toString(),
+                decimals: transaction.assetConfig.decimals,
+                total: transaction.assetConfig.total,
+                defaultFrozen: transaction.assetConfig.defaultFrozen,
+                metadataHash: transaction.assetConfig.assetMetadataHash,
+                ...(transaction.assetConfig.unitName
+                  ? {
+                    unitName: transaction.assetConfig.unitName,
+                    unitNameB64: Buffer.from(transaction.assetConfig.unitName).toString('base64'),
+                  }
+                  : undefined),
+                ...(transaction.assetConfig.assetName
+                  ? {
+                    name: transaction.assetConfig.assetName,
+                    nameB64: Buffer.from(transaction.assetConfig.assetName).toString('base64'),
+                  }
+                  : undefined),
+                ...(transaction.assetConfig.assetURL
+                  ? {
+                    url: transaction.assetConfig.assetURL,
+                    urlB64: Buffer.from(transaction.assetConfig.assetURL).toString('base64'),
+                  }
+                  : undefined),
+                manager: transaction.assetConfig.manager?.toString(),
+                reserve: transaction.assetConfig.reserve?.toString(),
+                clawback: transaction.assetConfig.clawback?.toString(),
+                freeze: transaction.assetConfig.freeze?.toString(),
+              })
+              : // In algosdk, transaction.assetConfig is always defined, even if it's an asset destroy transaction
+              // If any field is truthy, it's an asset update, otherwise, it's an asset destroy
+              transaction.assetConfig.manager ||
+                transaction.assetConfig.reserve ||
+                transaction.assetConfig.clawback ||
+                transaction.assetConfig.freeze ||
+                transaction.assetConfig.unitName ||
+                transaction.assetConfig.assetName ||
+                transaction.assetConfig.assetURL ||
+                transaction.assetConfig.assetMetadataHash
                 ? new algosdk.indexerModels.AssetParams({
-                    creator: transaction.sender.toString(),
-                    decimals: transaction.assetConfig.decimals,
-                    total: transaction.assetConfig.total,
-                    defaultFrozen: transaction.assetConfig.defaultFrozen,
-                    metadataHash: transaction.assetConfig.assetMetadataHash,
-                    ...(transaction.assetConfig.unitName
-                      ? {
-                          unitName: transaction.assetConfig.unitName,
-                          unitNameB64: Buffer.from(transaction.assetConfig.unitName).toString('base64'),
-                        }
-                      : undefined),
-                    ...(transaction.assetConfig.assetName
-                      ? {
-                          name: transaction.assetConfig.assetName,
-                          nameB64: Buffer.from(transaction.assetConfig.assetName).toString('base64'),
-                        }
-                      : undefined),
-                    ...(transaction.assetConfig.assetURL
-                      ? {
-                          url: transaction.assetConfig.assetURL,
-                          urlB64: Buffer.from(transaction.assetConfig.assetURL).toString('base64'),
-                        }
-                      : undefined),
-                    manager: transaction.assetConfig.manager?.toString(),
-                    reserve: transaction.assetConfig.reserve?.toString(),
-                    clawback: transaction.assetConfig.clawback?.toString(),
-                    freeze: transaction.assetConfig.freeze?.toString(),
-                  })
-                : // In algosdk, transaction.assetConfig is always defined, even if it's an asset destroy transaction
-                  // If any field is truthy, it's an asset update, otherwise, it's an asset destroy
-                  transaction.assetConfig.manager ||
-                    transaction.assetConfig.reserve ||
-                    transaction.assetConfig.clawback ||
-                    transaction.assetConfig.freeze ||
-                    transaction.assetConfig.unitName ||
-                    transaction.assetConfig.assetName ||
-                    transaction.assetConfig.assetURL ||
-                    transaction.assetConfig.assetMetadataHash
-                  ? new algosdk.indexerModels.AssetParams({
-                      manager: transaction.assetConfig.manager?.toString(),
-                      reserve: transaction.assetConfig.reserve?.toString(),
-                      clawback: transaction.assetConfig.clawback?.toString(),
-                      freeze: transaction.assetConfig.freeze?.toString(),
-                      // These parameters are required in the indexer type so setting to empty values
-                      creator: '',
-                      decimals: 0,
-                      total: 0,
-                    })
-                  : undefined,
-            }),
-          }
+                  manager: transaction.assetConfig.manager?.toString(),
+                  reserve: transaction.assetConfig.reserve?.toString(),
+                  clawback: transaction.assetConfig.clawback?.toString(),
+                  freeze: transaction.assetConfig.freeze?.toString(),
+                  // These parameters are required in the indexer type so setting to empty values
+                  creator: '',
+                  decimals: 0,
+                  total: 0,
+                })
+                : undefined,
+          }),
+        }
         : undefined),
       ...(transaction.type === TransactionType.axfer
         ? {
-            assetTransferTransaction: new algosdk.indexerModels.TransactionAssetTransfer({
-              assetId: transaction.assetTransfer!.assetIndex,
-              amount: transaction.assetTransfer!.amount,
-              receiver: transaction.assetTransfer!.receiver.toString(),
-              sender: transaction.assetTransfer!.assetSender ? transaction.assetTransfer!.assetSender.toString() : undefined,
-              closeAmount: assetCloseAmount,
-              closeTo: transaction.assetTransfer!.closeRemainderTo ? transaction.assetTransfer!.closeRemainderTo.toString() : undefined,
-            }),
-          }
+          assetTransferTransaction: new algosdk.indexerModels.TransactionAssetTransfer({
+            assetId: transaction.assetTransfer!.assetIndex,
+            amount: transaction.assetTransfer!.amount,
+            receiver: transaction.assetTransfer!.receiver.toString(),
+            sender: transaction.assetTransfer!.assetSender ? transaction.assetTransfer!.assetSender.toString() : undefined,
+            closeAmount: assetCloseAmount,
+            closeTo: transaction.assetTransfer!.closeRemainderTo ? transaction.assetTransfer!.closeRemainderTo.toString() : undefined,
+          }),
+        }
         : undefined),
       ...(transaction.type === TransactionType.afrz
         ? {
-            assetFreezeTransaction: new algosdk.indexerModels.TransactionAssetFreeze({
-              assetId: transaction.assetFreeze!.assetIndex,
-              newFreezeStatus: transaction.assetFreeze!.frozen,
-              address: transaction.assetFreeze!.freezeAccount.toString(),
-            }),
-          }
+          assetFreezeTransaction: new algosdk.indexerModels.TransactionAssetFreeze({
+            assetId: transaction.assetFreeze!.assetIndex,
+            newFreezeStatus: transaction.assetFreeze!.frozen,
+            address: transaction.assetFreeze!.freezeAccount.toString(),
+          }),
+        }
         : undefined),
       ...(transaction.type === TransactionType.appl
         ? {
-            applicationTransaction: new algosdk.indexerModels.TransactionApplication({
-              applicationId: transaction.applicationCall!.appIndex ?? 0,
-              approvalProgram: transaction.applicationCall!.approvalProgram,
-              clearStateProgram: transaction.applicationCall!.clearProgram,
-              onCompletion: algodOnCompleteToIndexerOnComplete(transaction.applicationCall!.onComplete),
-              applicationArgs: transaction.applicationCall!.appArgs.map((a) => a),
-              foreignApps: transaction.applicationCall!.foreignApps.map((a) => a),
-              foreignAssets: transaction.applicationCall!.foreignAssets.map((a) => a),
-              ...(transaction.applicationCall
-                ? {
-                    globalStateSchema: new algosdk.indexerModels.StateSchema({
-                      numByteSlice: transaction.applicationCall.numGlobalByteSlices,
-                      numUint: transaction.applicationCall.numGlobalInts,
-                    }),
-                  }
-                : undefined),
-              ...(transaction.applicationCall
-                ? {
-                    localStateSchema: new algosdk.indexerModels.StateSchema({
-                      numByteSlice: transaction.applicationCall.numLocalByteSlices,
-                      numUint: transaction.applicationCall.numLocalInts,
-                    }),
-                  }
-                : undefined),
-              accounts: transaction.applicationCall!.accounts.map((a) => a),
-              extraProgramPages: transaction.applicationCall!.extraPages,
-            }),
-          }
+          applicationTransaction: new algosdk.indexerModels.TransactionApplication({
+            applicationId: transaction.applicationCall!.appIndex ?? 0,
+            approvalProgram: transaction.applicationCall!.approvalProgram,
+            clearStateProgram: transaction.applicationCall!.clearProgram,
+            onCompletion: algodOnCompleteToIndexerOnComplete(transaction.applicationCall!.onComplete),
+            applicationArgs: transaction.applicationCall!.appArgs.map((a) => a),
+            foreignApps: transaction.applicationCall!.foreignApps.map((a) => a),
+            foreignAssets: transaction.applicationCall!.foreignAssets.map((a) => a),
+            ...(transaction.applicationCall
+              ? {
+                globalStateSchema: new algosdk.indexerModels.StateSchema({
+                  numByteSlice: transaction.applicationCall.numGlobalByteSlices,
+                  numUint: transaction.applicationCall.numGlobalInts,
+                }),
+              }
+              : undefined),
+            ...(transaction.applicationCall
+              ? {
+                localStateSchema: new algosdk.indexerModels.StateSchema({
+                  numByteSlice: transaction.applicationCall.numLocalByteSlices,
+                  numUint: transaction.applicationCall.numLocalInts,
+                }),
+              }
+              : undefined),
+            accounts: transaction.applicationCall!.accounts.map((a) => a),
+            extraProgramPages: transaction.applicationCall!.extraPages,
+          }),
+        }
         : undefined),
       ...(transaction.type === TransactionType.pay
         ? {
-            paymentTransaction: new algosdk.indexerModels.TransactionPayment({
-              amount: transaction.payment!.amount,
-              receiver: transaction.payment!.receiver.toString(),
-              closeAmount: closeAmount,
-              closeRemainderTo: transaction.payment!.closeRemainderTo?.toString(),
-            }),
-          }
+          paymentTransaction: new algosdk.indexerModels.TransactionPayment({
+            amount: transaction.payment!.amount,
+            receiver: transaction.payment!.receiver.toString(),
+            closeAmount: closeAmount,
+            closeRemainderTo: transaction.payment!.closeRemainderTo?.toString(),
+          }),
+        }
         : undefined),
       ...(transaction.type === TransactionType.keyreg
         ? {
-            keyregTransaction: new algosdk.indexerModels.TransactionKeyreg({
-              nonParticipation: transaction.keyreg!.nonParticipation,
-              selectionParticipationKey: transaction.keyreg!.selectionKey,
-              stateProofKey: transaction.keyreg!.stateProofKey,
-              voteFirstValid: transaction.keyreg!.voteFirst,
-              voteKeyDilution: transaction.keyreg!.voteKeyDilution,
-              voteLastValid: transaction.keyreg!.voteLast,
-              voteParticipationKey: transaction.keyreg!.voteKey,
-            }),
-          }
+          keyregTransaction: new algosdk.indexerModels.TransactionKeyreg({
+            nonParticipation: transaction.keyreg!.nonParticipation,
+            selectionParticipationKey: transaction.keyreg!.selectionKey,
+            stateProofKey: transaction.keyreg!.stateProofKey,
+            voteFirstValid: transaction.keyreg!.voteFirst,
+            voteKeyDilution: transaction.keyreg!.voteKeyDilution,
+            voteLastValid: transaction.keyreg!.voteLast,
+            voteParticipationKey: transaction.keyreg!.voteKey,
+          }),
+        }
         : undefined),
       ...(transaction.type === TransactionType.stpf
         ? {
-            stateProofTransaction: new algosdk.indexerModels.TransactionStateProof({
-              stateProof: new algosdk.indexerModels.StateProofFields({
-                partProofs: transaction.stateProof!.stateProof?.partProofs
-                  ? algodMerkleArrayProofToIndexerMerkleArrayProof(transaction.stateProof!.stateProof.partProofs)
-                  : undefined,
-                positionsToReveal: transaction.stateProof!.stateProof?.positionsToReveal,
-                saltVersion: transaction.stateProof!.stateProof?.merkleSignatureSaltVersion,
-                sigCommit: transaction.stateProof!.stateProof?.sigCommit,
-                sigProofs: transaction.stateProof!.stateProof?.sigProofs
-                  ? algodMerkleArrayProofToIndexerMerkleArrayProof(transaction.stateProof!.stateProof.sigProofs)
-                  : undefined,
-                signedWeight: transaction.stateProof!.stateProof?.signedWeight,
-                reveals: transaction.stateProof!.stateProof?.reveals
-                  ? Array.from(transaction.stateProof!.stateProof?.reveals.entries()).map(([position, reveal]) => {
-                      return new algosdk.indexerModels.StateProofReveal({
-                        sigSlot: new algosdk.indexerModels.StateProofSigSlot({
-                          lowerSigWeight: reveal.sigslot.l,
-                          signature: new algosdk.indexerModels.StateProofSignature({
-                            merkleArrayIndex: reveal.sigslot.sig.vectorCommitmentIndex,
-                            falconSignature: Buffer.from(reveal.sigslot.sig.signature),
-                            proof: algodMerkleArrayProofToIndexerMerkleArrayProof(reveal.sigslot.sig.proof),
-                            verifyingKey: reveal.sigslot.sig.verifyingKey.publicKey,
-                          }),
-                        }),
-                        position: position,
-                        participant: new algosdk.indexerModels.StateProofParticipant({
-                          weight: reveal.participant.weight,
-                          verifier: new algosdk.indexerModels.StateProofVerifier({
-                            keyLifetime: reveal.participant.pk.keyLifetime,
-                            commitment: reveal.participant.pk.commitment,
-                          }),
-                        }),
-                      })
-                    })
-                  : undefined,
-              }),
-              message: new algosdk.indexerModels.IndexerStateProofMessage({
-                blockHeadersCommitment: Buffer.from(transaction.stateProof!.message!.blockHeadersCommitment),
-                firstAttestedRound: transaction.stateProof!.message!.firstAttestedRound,
-                latestAttestedRound: transaction.stateProof!.message!.lastAttestedRound,
-                lnProvenWeight: transaction.stateProof!.message!.lnProvenWeight,
-                votersCommitment: Buffer.from(transaction.stateProof!.message!.votersCommitment),
-              }),
-              stateProofType: Number(transaction.stateProof!.stateProofType ?? 0),
+          stateProofTransaction: new algosdk.indexerModels.TransactionStateProof({
+            stateProof: new algosdk.indexerModels.StateProofFields({
+              partProofs: transaction.stateProof!.stateProof?.partProofs
+                ? algodMerkleArrayProofToIndexerMerkleArrayProof(transaction.stateProof!.stateProof.partProofs)
+                : undefined,
+              positionsToReveal: transaction.stateProof!.stateProof?.positionsToReveal,
+              saltVersion: transaction.stateProof!.stateProof?.merkleSignatureSaltVersion,
+              sigCommit: transaction.stateProof!.stateProof?.sigCommit,
+              sigProofs: transaction.stateProof!.stateProof?.sigProofs
+                ? algodMerkleArrayProofToIndexerMerkleArrayProof(transaction.stateProof!.stateProof.sigProofs)
+                : undefined,
+              signedWeight: transaction.stateProof!.stateProof?.signedWeight,
+              reveals: transaction.stateProof!.stateProof?.reveals
+                ? Array.from(transaction.stateProof!.stateProof?.reveals.entries()).map(([position, reveal]) => {
+                  return new algosdk.indexerModels.StateProofReveal({
+                    sigSlot: new algosdk.indexerModels.StateProofSigSlot({
+                      lowerSigWeight: reveal.sigslot.l,
+                      signature: new algosdk.indexerModels.StateProofSignature({
+                        merkleArrayIndex: reveal.sigslot.sig.vectorCommitmentIndex,
+                        falconSignature: Buffer.from(reveal.sigslot.sig.signature),
+                        proof: algodMerkleArrayProofToIndexerMerkleArrayProof(reveal.sigslot.sig.proof),
+                        verifyingKey: reveal.sigslot.sig.verifyingKey.publicKey,
+                      }),
+                    }),
+                    position: position,
+                    participant: new algosdk.indexerModels.StateProofParticipant({
+                      weight: reveal.participant.weight,
+                      verifier: new algosdk.indexerModels.StateProofVerifier({
+                        keyLifetime: reveal.participant.pk.keyLifetime,
+                        commitment: reveal.participant.pk.commitment,
+                      }),
+                    }),
+                  })
+                })
+                : undefined,
             }),
-          }
+            message: new algosdk.indexerModels.IndexerStateProofMessage({
+              blockHeadersCommitment: Buffer.from(transaction.stateProof!.message!.blockHeadersCommitment),
+              firstAttestedRound: transaction.stateProof!.message!.firstAttestedRound,
+              latestAttestedRound: transaction.stateProof!.message!.lastAttestedRound,
+              lnProvenWeight: transaction.stateProof!.message!.lnProvenWeight,
+              votersCommitment: Buffer.from(transaction.stateProof!.message!.votersCommitment),
+            }),
+            stateProofType: Number(transaction.stateProof!.stateProofType ?? 0),
+          }),
+        }
         : undefined),
       ...(transaction.type === TransactionType.hb && transaction.heartbeat
         ? {
-            heartbeatTransaction: new algosdk.indexerModels.TransactionHeartbeat({
-              hbAddress: transaction.heartbeat.address.toString(),
-              hbKeyDilution: transaction.heartbeat.keyDilution,
-              hbProof: new algosdk.indexerModels.HbProofFields({
-                hbPk: transaction.heartbeat.proof.pk,
-                hbPk1sig: transaction.heartbeat.proof.pk1Sig,
-                hbPk2: transaction.heartbeat.proof.pk2,
-                hbPk2sig: transaction.heartbeat.proof.pk2Sig,
-                hbSig: transaction.heartbeat.proof.sig,
-              }),
-              hbSeed: transaction.heartbeat.seed,
-              hbVoteId: transaction.heartbeat.voteID,
+          heartbeatTransaction: new algosdk.indexerModels.TransactionHeartbeat({
+            hbAddress: transaction.heartbeat.address.toString(),
+            hbKeyDilution: transaction.heartbeat.keyDilution,
+            hbProof: new algosdk.indexerModels.HbProofFields({
+              hbPk: transaction.heartbeat.proof.pk,
+              hbPk1sig: transaction.heartbeat.proof.pk1Sig,
+              hbPk2: transaction.heartbeat.proof.pk2,
+              hbPk2sig: transaction.heartbeat.proof.pk2Sig,
+              hbSig: transaction.heartbeat.proof.sig,
             }),
-          }
+            hbSeed: transaction.heartbeat.seed,
+            hbVoteId: transaction.heartbeat.voteID,
+          }),
+        }
         : undefined),
       firstValid: transaction.firstValid,
       lastValid: transaction.lastValid,
@@ -421,35 +466,20 @@ export function getIndexerTransactionFromAlgodTransaction(t: TransactionInBlock,
       }),
       ...(signedTxnWithAD.signedTxn.sig || signedTxnWithAD.signedTxn.lsig || signedTxnWithAD.signedTxn.msig
         ? {
-            signature: new algosdk.indexerModels.TransactionSignature({
-              sig: signedTxnWithAD.signedTxn.sig ? Buffer.from(signedTxnWithAD.signedTxn.sig).toString('base64') : undefined,
-              logicsig: signedTxnWithAD.signedTxn.lsig
-                ? new algosdk.indexerModels.TransactionSignatureLogicsig({
-                    logic: Buffer.from(signedTxnWithAD.signedTxn.lsig.logic).toString('base64'),
-                    args: signedTxnWithAD.signedTxn.lsig.args,
-                    signature: signedTxnWithAD.signedTxn.lsig.sig
-                      ? Buffer.from(signedTxnWithAD.signedTxn.lsig.sig).toString('base64')
-                      : undefined,
-                    multisigSignature: signedTxnWithAD.signedTxn.lsig.msig
-                      ? new algosdk.indexerModels.TransactionSignatureMultisig({
-                          version: signedTxnWithAD.signedTxn.lsig.msig.v,
-                          threshold: signedTxnWithAD.signedTxn.lsig.msig.thr,
-                          subsignature: signedTxnWithAD.signedTxn.lsig.msig.subsig.map(
-                            (s) =>
-                              new algosdk.indexerModels.TransactionSignatureMultisigSubsignature({
-                                publicKey: Buffer.from(s.pk).toString('base64'),
-                                signature: s.s ? Buffer.from(s.s).toString('base64') : undefined,
-                              }),
-                          ),
-                        })
-                      : undefined,
-                  })
-                : undefined,
-              multisig: signedTxnWithAD.signedTxn.msig
-                ? new algosdk.indexerModels.TransactionSignatureMultisig({
-                    version: signedTxnWithAD.signedTxn.msig.v,
-                    threshold: signedTxnWithAD.signedTxn.msig.thr,
-                    subsignature: signedTxnWithAD.signedTxn.msig.subsig.map(
+          signature: new algosdk.indexerModels.TransactionSignature({
+            sig: signedTxnWithAD.signedTxn.sig ? Buffer.from(signedTxnWithAD.signedTxn.sig).toString('base64') : undefined,
+            logicsig: signedTxnWithAD.signedTxn.lsig
+              ? new algosdk.indexerModels.TransactionSignatureLogicsig({
+                logic: Buffer.from(signedTxnWithAD.signedTxn.lsig.logic).toString('base64'),
+                args: signedTxnWithAD.signedTxn.lsig.args,
+                signature: signedTxnWithAD.signedTxn.lsig.sig
+                  ? Buffer.from(signedTxnWithAD.signedTxn.lsig.sig).toString('base64')
+                  : undefined,
+                multisigSignature: signedTxnWithAD.signedTxn.lsig.msig
+                  ? new algosdk.indexerModels.TransactionSignatureMultisig({
+                    version: signedTxnWithAD.signedTxn.lsig.msig.v,
+                    threshold: signedTxnWithAD.signedTxn.lsig.msig.thr,
+                    subsignature: signedTxnWithAD.signedTxn.lsig.msig.subsig.map(
                       (s) =>
                         new algosdk.indexerModels.TransactionSignatureMultisigSubsignature({
                           publicKey: Buffer.from(s.pk).toString('base64'),
@@ -457,9 +487,24 @@ export function getIndexerTransactionFromAlgodTransaction(t: TransactionInBlock,
                         }),
                     ),
                   })
-                : undefined,
-            }),
-          }
+                  : undefined,
+              })
+              : undefined,
+            multisig: signedTxnWithAD.signedTxn.msig
+              ? new algosdk.indexerModels.TransactionSignatureMultisig({
+                version: signedTxnWithAD.signedTxn.msig.v,
+                threshold: signedTxnWithAD.signedTxn.msig.thr,
+                subsignature: signedTxnWithAD.signedTxn.msig.subsig.map(
+                  (s) =>
+                    new algosdk.indexerModels.TransactionSignatureMultisigSubsignature({
+                      publicKey: Buffer.from(s.pk).toString('base64'),
+                      signature: s.s ? Buffer.from(s.s).toString('base64') : undefined,
+                    }),
+                ),
+              })
+              : undefined,
+          }),
+        }
         : undefined),
       logs: signedTxnWithAD.applyData.evalDelta?.logs,
       closeRewards: closeRewards,
@@ -467,47 +512,47 @@ export function getIndexerTransactionFromAlgodTransaction(t: TransactionInBlock,
       senderRewards: senderRewards,
       globalStateDelta: signedTxnWithAD.applyData.evalDelta?.globalDelta
         ? Array.from(signedTxnWithAD.applyData.evalDelta?.globalDelta.entries()).map(
-            ([key, value]) =>
-              new algosdk.indexerModels.EvalDeltaKeyValue({
-                key: Buffer.from(key).toString('base64'),
-                value: new algosdk.indexerModels.EvalDelta({
-                  action: value.action,
-                  ...(value.action === 2
-                    ? {
-                        uint: value.uint,
-                      }
-                    : {
-                        bytes: value.bytes ? Buffer.from(value.bytes).toString('base64') : undefined,
-                        uint: value.uint,
-                      }),
-                }),
+          ([key, value]) =>
+            new algosdk.indexerModels.EvalDeltaKeyValue({
+              key: Buffer.from(key).toString('base64'),
+              value: new algosdk.indexerModels.EvalDelta({
+                action: value.action,
+                ...(value.action === 2
+                  ? {
+                    uint: value.uint,
+                  }
+                  : {
+                    bytes: value.bytes ? Buffer.from(value.bytes).toString('base64') : undefined,
+                    uint: value.uint,
+                  }),
               }),
-          )
+            }),
+        )
         : undefined,
       localStateDelta: signedTxnWithAD.applyData.evalDelta?.localDeltas
         ? Array.from(signedTxnWithAD.applyData.evalDelta?.localDeltas.entries()).map(([addressIndex, delta]) => {
-            const addresses = [transaction.sender.toString(), ...(transaction.applicationCall?.accounts?.map((a) => a.toString()) || [])]
-            return new algosdk.indexerModels.AccountStateDelta({
-              address: addresses[Number(addressIndex)],
-              delta: Array.from(delta.entries()).map(
-                ([key, value]) =>
-                  new algosdk.indexerModels.EvalDeltaKeyValue({
-                    key: Buffer.from(key).toString('base64'),
-                    value: new algosdk.indexerModels.EvalDelta({
-                      action: value.action,
-                      ...(value.action === 2
-                        ? {
-                            uint: value.uint,
-                          }
-                        : {
-                            bytes: value.bytes ? Buffer.from(value.bytes).toString('base64') : undefined,
-                            uint: value.uint,
-                          }),
-                    }),
+          const addresses = [transaction.sender.toString(), ...(transaction.applicationCall?.accounts?.map((a) => a.toString()) || [])]
+          return new algosdk.indexerModels.AccountStateDelta({
+            address: addresses[Number(addressIndex)],
+            delta: Array.from(delta.entries()).map(
+              ([key, value]) =>
+                new algosdk.indexerModels.EvalDeltaKeyValue({
+                  key: Buffer.from(key).toString('base64'),
+                  value: new algosdk.indexerModels.EvalDelta({
+                    action: value.action,
+                    ...(value.action === 2
+                      ? {
+                        uint: value.uint,
+                      }
+                      : {
+                        bytes: value.bytes ? Buffer.from(value.bytes).toString('base64') : undefined,
+                        uint: value.uint,
+                      }),
                   }),
-              ),
-            })
+                }),
+            ),
           })
+        })
         : undefined,
     })
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -522,8 +567,8 @@ function algodMerkleArrayProofToIndexerMerkleArrayProof(proof: algosdk.MerkleArr
   return new algosdk.indexerModels.MerkleArrayProof({
     hashFactory: proof.hashFactory
       ? new algosdk.indexerModels.HashFactory({
-          hashType: proof.hashFactory.hashType,
-        })
+        hashType: proof.hashFactory.hashType,
+      })
       : undefined,
     path: proof.path,
     treeDepth: proof.treeDepth,
@@ -588,28 +633,28 @@ export function blockResponseToBlockMetadata(blockResponse: algosdk.modelsv2.Blo
     proposer: block.header.proposer?.toString(),
     ...(block.header.upgradeVote
       ? {
-          upgradeVote: {
-            upgradeApprove: block.header.upgradeVote.upgradeApprove,
-            upgradeDelay: block.header.upgradeVote.upgradeDelay,
-            upgradePropose: block.header.upgradeVote.upgradePropose,
-          },
-        }
+        upgradeVote: {
+          upgradeApprove: block.header.upgradeVote.upgradeApprove,
+          upgradeDelay: block.header.upgradeVote.upgradeDelay,
+          upgradePropose: block.header.upgradeVote.upgradePropose,
+        },
+      }
       : undefined),
     ...(block.header.participationUpdates
       ? {
-          participationUpdates: {
-            absentParticipationAccounts: block.header.participationUpdates.absentParticipationAccounts.map((addr) => addr.toString()),
-            expiredParticipationAccounts: block.header.participationUpdates.expiredParticipationAccounts.map((addr) => addr.toString()),
-          },
-        }
+        participationUpdates: {
+          absentParticipationAccounts: block.header.participationUpdates.absentParticipationAccounts.map((addr) => addr.toString()),
+          expiredParticipationAccounts: block.header.participationUpdates.expiredParticipationAccounts.map((addr) => addr.toString()),
+        },
+      }
       : undefined),
     stateProofTracking: block.header.stateproofTracking
       ? Array.from(block.header.stateproofTracking.entries()).map(([key, value]) => ({
-          nextRound: value.stateProofNextRound,
-          onlineTotalWeight: value.stateProofOnlineTotalWeight ?? 0n,
-          type: Number(key),
-          votersCommitment: Buffer.from(value.stateProofVotersCommitment).toString('base64'),
-        }))
+        nextRound: value.stateProofNextRound,
+        onlineTotalWeight: value.stateProofOnlineTotalWeight ?? 0n,
+        type: Number(key),
+        votersCommitment: Buffer.from(value.stateProofVotersCommitment).toString('base64'),
+      }))
       : undefined,
   }
 }
@@ -646,29 +691,29 @@ export function extractBalanceChangesFromBlockTransaction(signedTxnWithAD: Signe
       },
       ...(transaction.payment?.receiver
         ? [
-            {
-              address: transaction.payment.receiver.toString(),
-              amount: transaction.payment.amount,
-              roles: [BalanceChangeRole.Receiver],
-              assetId: 0n,
-            },
-          ]
+          {
+            address: transaction.payment.receiver.toString(),
+            amount: transaction.payment.amount,
+            roles: [BalanceChangeRole.Receiver],
+            assetId: 0n,
+          },
+        ]
         : []),
       ...(transaction.payment?.closeRemainderTo
         ? [
-            {
-              address: transaction.payment.closeRemainderTo.toString(),
-              amount: signedTxnWithAD.applyData.closingAmount ?? 0n,
-              roles: [BalanceChangeRole.CloseTo],
-              assetId: 0n,
-            },
-            {
-              address: transaction.sender.toString(),
-              amount: -1n * (signedTxnWithAD.applyData.closingAmount ?? 0n),
-              roles: [BalanceChangeRole.Sender],
-              assetId: 0n,
-            },
-          ]
+          {
+            address: transaction.payment.closeRemainderTo.toString(),
+            amount: signedTxnWithAD.applyData.closingAmount ?? 0n,
+            roles: [BalanceChangeRole.CloseTo],
+            assetId: 0n,
+          },
+          {
+            address: transaction.sender.toString(),
+            amount: -1n * (signedTxnWithAD.applyData.closingAmount ?? 0n),
+            roles: [BalanceChangeRole.Sender],
+            assetId: 0n,
+          },
+        ]
         : []),
     )
   }
@@ -683,29 +728,29 @@ export function extractBalanceChangesFromBlockTransaction(signedTxnWithAD: Signe
       },
       ...(transaction.assetTransfer!.receiver
         ? [
-            {
-              address: transaction.assetTransfer!.receiver.toString(),
-              assetId: transaction.assetTransfer!.assetIndex,
-              amount: transaction.assetTransfer!.amount,
-              roles: [BalanceChangeRole.Receiver],
-            },
-          ]
+          {
+            address: transaction.assetTransfer!.receiver.toString(),
+            assetId: transaction.assetTransfer!.assetIndex,
+            amount: transaction.assetTransfer!.amount,
+            roles: [BalanceChangeRole.Receiver],
+          },
+        ]
         : []),
       ...(transaction.assetTransfer!.closeRemainderTo
         ? [
-            {
-              address: transaction.assetTransfer!.closeRemainderTo.toString(),
-              assetId: transaction.assetTransfer!.assetIndex,
-              amount: signedTxnWithAD.applyData.assetClosingAmount ?? 0n,
-              roles: [BalanceChangeRole.CloseTo],
-            },
-            {
-              address: transaction.sender.toString(),
-              assetId: transaction.assetTransfer!.assetIndex,
-              amount: -1n * (signedTxnWithAD.applyData.assetClosingAmount ?? 0n),
-              roles: [BalanceChangeRole.Sender],
-            },
-          ]
+          {
+            address: transaction.assetTransfer!.closeRemainderTo.toString(),
+            assetId: transaction.assetTransfer!.assetIndex,
+            amount: signedTxnWithAD.applyData.assetClosingAmount ?? 0n,
+            roles: [BalanceChangeRole.CloseTo],
+          },
+          {
+            address: transaction.sender.toString(),
+            assetId: transaction.assetTransfer!.assetIndex,
+            amount: -1n * (signedTxnWithAD.applyData.assetClosingAmount ?? 0n),
+            roles: [BalanceChangeRole.Sender],
+          },
+        ]
         : []),
     )
   }
@@ -778,19 +823,19 @@ export function extractBalanceChangesFromIndexerTransaction(transaction: algosdk
       },
       ...(pay.closeAmount
         ? [
-            {
-              address: pay.closeRemainderTo!,
-              amount: pay.closeAmount,
-              roles: [BalanceChangeRole.CloseTo],
-              assetId: 0n,
-            },
-            {
-              address: transaction.sender,
-              amount: -1n * pay.closeAmount,
-              roles: [BalanceChangeRole.Sender],
-              assetId: 0n,
-            },
-          ]
+          {
+            address: pay.closeRemainderTo!,
+            amount: pay.closeAmount,
+            roles: [BalanceChangeRole.CloseTo],
+            assetId: 0n,
+          },
+          {
+            address: transaction.sender,
+            amount: -1n * pay.closeAmount,
+            roles: [BalanceChangeRole.Sender],
+            assetId: 0n,
+          },
+        ]
         : []),
     )
   }
@@ -812,19 +857,19 @@ export function extractBalanceChangesFromIndexerTransaction(transaction: algosdk
       },
       ...(axfer.closeAmount && axfer.closeTo
         ? [
-            {
-              address: axfer.closeTo,
-              assetId: axfer.assetId,
-              amount: axfer.closeAmount,
-              roles: [BalanceChangeRole.CloseTo],
-            },
-            {
-              address: axfer.sender ?? transaction.sender,
-              assetId: axfer.assetId,
-              amount: -1n * axfer.closeAmount,
-              roles: [BalanceChangeRole.Sender],
-            },
-          ]
+          {
+            address: axfer.closeTo,
+            assetId: axfer.assetId,
+            amount: axfer.closeAmount,
+            roles: [BalanceChangeRole.CloseTo],
+          },
+          {
+            address: axfer.sender ?? transaction.sender,
+            assetId: axfer.assetId,
+            amount: -1n * axfer.closeAmount,
+            roles: [BalanceChangeRole.Sender],
+          },
+        ]
         : []),
     )
   }
